实现了APACE实时读取深度点云用TSDF建图的功能，可以在线飞行，但还不能重规划，容易撞障碍物

+++ APACE/utils/src/transformer/transformer.cpp
@@ -1,6 +1,7 @@
 #include "transformer/transformer.h"
 
-Transformer::Transformer(ros::NodeHandle &nh) : world_frame_("world"), sensor_frame_("camera") {
+Transformer::Transformer(ros::NodeHandle &nh) : world_frame_("world"), sensor_frame_("camera")
+{
   double timestamp_tolerance_duration;
   nh.param("transformer/world_frame", world_frame_, world_frame_);
   nh.param("transformer/sensor_frame", sensor_frame_, sensor_frame_);
@@ -9,18 +10,26 @@ Transformer::Transformer(ros::NodeHandle &nh) : world_frame_("world"), sensor_fr
 
   std::string pose_topic_type;
   nh.param("transformer/pose_topic_type", pose_topic_type, std::string(""));
-  if (pose_topic_type == "pose") {
+  if (pose_topic_type == "pose")
+  {
     config_.pose_topic_type_ = Config::POSE_TYPE::POSE;
-  } else if (pose_topic_type == "odometry") {
+  }
+  else if (pose_topic_type == "odometry")
+  {
     config_.pose_topic_type_ = Config::POSE_TYPE::ODOM;
-  } else if (pose_topic_type == "transform") {
+  }
+  else if (pose_topic_type == "transform")
+  {
     config_.pose_topic_type_ = Config::POSE_TYPE::TRANSFORM;
-  } else {
+  }
+  else
+  {
     config_.pose_topic_type_ = Config::POSE_TYPE::UNKNOWN;
   }
 
   timestamp_tolerance_ = ros::Duration(timestamp_tolerance_duration);
-  switch (config_.pose_topic_type_) {
+  switch (config_.pose_topic_type_)
+  {
   case Config::POSE_TYPE::POSE:
     transform_sub_ =
         nh.subscribe("/transformer/sensor_pose_topic", 100, &Transformer::poseCallback, this);
@@ -40,7 +49,8 @@ Transformer::Transformer(ros::NodeHandle &nh) : world_frame_("world"), sensor_fr
     break;
   }
 
-  if (verbose_) {
+  if (verbose_)
+  {
     ROS_INFO("[Transformer] Initialized transformer with world frame: %s, sensor frame: %s",
              world_frame_.c_str(), sensor_frame_.c_str());
     ROS_INFO("[Transformer] Subscribe from transformer topic: %s",
@@ -48,34 +58,43 @@ Transformer::Transformer(ros::NodeHandle &nh) : world_frame_("world"), sensor_fr
   }
 }
 
-bool Transformer::lookupTransform(const ros::Time &timestamp, Transformation &transform) {
-  if (transform_queue_.empty()) {
+bool Transformer::lookupTransform(const ros::Time &timestamp, Transformation &transform)
+{
+  if (transform_queue_.empty())
+  {
     if (verbose_)
-      ROS_WARN_STREAM_THROTTLE(1, "[Transformer] No match found for transform timestamp: "
-                                      << timestamp << " as transform queue is empty.");
+      ROS_WARN_STREAM_THROTTLE(1, "[Transformer] No match found for transform timestamp: " << timestamp << " as transform queue is empty.");
     return false;
   }
 
   bool match_found = false;
   std::deque<TransformStampedMsgType>::reverse_iterator it = transform_queue_.rbegin();
-  for (; it != transform_queue_.rend(); ++it) {
-    if (it->header.stamp < timestamp) {
-      if (timestamp - it->header.stamp < timestamp_tolerance_) {
+  for (; it != transform_queue_.rend(); ++it)
+  {
+    if (it->header.stamp < timestamp)
+    {
+      if (timestamp - it->header.stamp < timestamp_tolerance_)
+      {
         match_found = true;
       }
       break;
     }
 
-    if (it->header.stamp - timestamp < timestamp_tolerance_) {
+    if (it->header.stamp - timestamp < timestamp_tolerance_)
+    {
       match_found = true;
       break;
     }
   }
 
-  if (match_found) {
+  if (match_found)
+  {
     tf::transformMsgToKindr(it->transform, &transform);
-  } else {
-    if (it == transform_queue_.rbegin() || it == transform_queue_.rend()) {
+  }
+  else
+  {
+    if (it == transform_queue_.rbegin() || it == transform_queue_.rend())
+    {
       if (verbose_)
         ROS_WARN_STREAM("[Transformer] No match found for transform timestamp: "
                         << timestamp << " Queue front: " << transform_queue_.front().header.stamp
@@ -113,7 +132,8 @@ bool Transformer::lookupTransform(const ros::Time &timestamp, Transformation &tr
   return true;
 }
 
-void Transformer::transformCallback(const TransformStampedMsgType &msg) {
+void Transformer::transformCallback(const TransformStampedMsgType &msg)
+{
   if (verbose_)
     ROS_INFO_THROTTLE(1.0, "[Transformer] Received transform, transformation queue size: %d",
                       transform_queue_.size());
@@ -121,7 +141,8 @@ void Transformer::transformCallback(const TransformStampedMsgType &msg) {
   transform_queue_.push_back(msg);
 }
 
-void Transformer::poseCallback(const PoseStampedMsgType &msg) {
+void Transformer::poseCallback(const PoseStampedMsgType &msg)
+{
   if (verbose_)
     ROS_INFO_THROTTLE(1.0, "[Transformer] Received transform, transformation queue size: %d",
                       transform_queue_.size());
@@ -138,10 +159,10 @@ void Transformer::poseCallback(const PoseStampedMsgType &msg) {
   transform_queue_.push_back(transform_msg);
 }
 
-void Transformer::odometryCallback(const OdometryMsgType &msg) {
+void Transformer::odometryCallback(const OdometryMsgType &msg)
+{
   if (verbose_)
-    ROS_INFO_THROTTLE(1.0, "[Transformer] Received transform, transformation queue size: %d",
-                      transform_queue_.size());
+    ROS_INFO_THROTTLE(1.0, "[Transformer] Received transform, transformation queue size: %d", static_cast<int>(transform_queue_.size()));
 
   TransformStampedMsgType transform_msg;
   transform_msg.header = msg.header;
